# Use ReadOnlySpan<byte> compiler optimization for static data