… ARMeilleure/Instructions/CryptoHelper.cs
… Ryujinx.Common/XXHash128.cs