https://vcsjones.dev/2019/02/01/csharp-readonly-span-bytes-static/
https://github.com/dotnet/runtime/issues/33780
Revival of #1283, but this time with a benchmark example to show the performance benefit.

Benchmark of BitUtils.CountLeadingZeroes:

|                     Method |     Mean |     Error |    StdDev |
|--------------------------- |---------:|----------:|----------:|
|  LeadingZeroCount_int_old | 4.605 μs | 0.0160 μs | 0.0149 μs |
| LeadingZeroCount_long_old | 9.432 μs | 0.0912 μs | 0.0761 μs |
|  LeadingZeroCount_int_new | 3.310 μs | 0.0082 μs | 0.0068 μs |
| LeadingZeroCount_long_new | 7.612 μs | 0.0172 μs | 0.0144 μs |

I mostly used the [LeadingZeroCount](https://github.com/dotnet/performance/blob/main/src/benchmarks/micro/libraries/System.Numerics.BitOperations/Perf_BitOperations.cs) benchmarks from dotnet/performance